Moving script detection to a shared location
